Split Codex triage runner modules
