+++ sidecar/src/triage/codex.ts
@@ -1,4 +1,4 @@
-import { mkdir, mkdtemp, rm, writeFile } from "node:fs/promises";
+import { mkdtemp, rm } from "node:fs/promises";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
 import { CodexAppServer, type JsonRpcNotification } from "../codex-app-server";
@@ -9,23 +9,11 @@ import {
 import { callHost } from "../host-bridge";
 import { errorDetails, logger } from "../logger";
 import { pickFastestCodexModel } from "../model-catalog";
-import { buildSystemPrompt, buildTickUserMessage } from "./prompts";
-import type {
-	TriageCandidate,
-	TriageProposal,
-	TriageTickParams,
-} from "./types";
+import { type CodexTriageDecision, parseCodexDecision } from "./codex-decision";
+import { buildCodexTriageInput } from "./codex-input";
+import type { TriageTickParams } from "./types";
 
-type TriageDecision = {
-	readonly proposals: TriageProposal[];
-	readonly skips: readonly { candidateId: string; reason: string }[];
-	readonly summary: string | null;
-};
-
-type CandidateBody = {
-	readonly candidate: TriageCandidate;
-	readonly body: string;
-};
+export { parseCodexDecision } from "./codex-decision";
 
 const CODEX_APPROVAL_POLICY = {
 	granular: {
@@ -47,7 +35,7 @@ export function startCodexTriageTick(
 		}): void;
 	},
 ): {
-	promise: Promise<TriageDecision & { cancelled: boolean }>;
+	promise: Promise<CodexTriageDecision & { cancelled: boolean }>;
 	abort(): void;
 } {
 	let server: CodexAppServer | null = null;
@@ -94,7 +82,7 @@ async function runCodexTriageTick(
 		setServer(server: CodexAppServer): void;
 		isCancelled(): boolean;
 	},
-): Promise<TriageDecision & { cancelled: boolean }> {
+): Promise<CodexTriageDecision & { cancelled: boolean }> {
 	const tickId = params.tickId || "(no-tick-id)";
 	const logTag = `triage[${tickId}]`;
 	const cwd = await mkdtemp(join(tmpdir(), "helmor-triage-codex-"));
@@ -102,9 +90,7 @@ async function runCodexTriageTick(
 
 	let server: CodexAppServer | null = null;
 	try {
-		const candidateBodies = await readCandidateBodies(params.candidates);
-		const prompt = buildCodexUserPrompt(params, candidateBodies);
-		const imagePaths = await writeCandidateImages(params.candidates, cwd);
+		const input = await buildCodexTriageInput(params, cwd);
 		hooks.emitProgress({
 			tool: "codex_triage",
 			argsPreview: `${params.candidates.length} candidates`,
@@ -176,10 +162,7 @@ async function runCodexTriageTick(
 				?.sendRequest("turn/start", {
 					threadId,
 					model,
-					input: [
-						{ type: "text", text: prompt, text_elements: [] },
-						...imagePaths.map((path) => ({ type: "localImage", path })),
-					],
+					input,
 				})
 				.catch(reject);
 		});
@@ -246,170 +229,6 @@ function handleCodexNotification(
 	}
 }
 
-async function readCandidateBodies(
-	candidates: readonly TriageCandidate[],
-): Promise<CandidateBody[]> {
-	return Promise.all(
-		candidates.map(async (candidate) => {
-			const isIm = candidate.source === "lark" || candidate.source === "slack";
-			const response = await callHost<{ body: string }>(
-				"triage.read_candidate",
-				{
-					candidateId: candidate.id,
-					...(isIm ? { tail: 80 } : {}),
-				},
-			);
-			return { candidate, body: response.body };
-		}),
-	);
-}
-
-async function writeCandidateImages(
-	candidates: readonly TriageCandidate[],
-	cwd: string,
-): Promise<string[]> {
-	const dir = join(cwd, "attachments");
-	const paths: string[] = [];
-	let index = 0;
-	for (const candidate of candidates) {
-		for (const attachment of candidate.attachments ?? []) {
-			if (!attachment.dataBase64 || !attachment.mimeType.startsWith("image/")) {
-				continue;
-			}
-			await mkdir(dir, { recursive: true });
-			index += 1;
-			const path = join(
-				dir,
-				`${index}-${safeFilePart(candidate.id)}.${extensionForMime(attachment.mimeType)}`,
-			);
-			await writeFile(path, Buffer.from(attachment.dataBase64, "base64"));
-			paths.push(path);
-		}
-	}
-	return paths;
-}
-
-function buildCodexUserPrompt(
-	params: TriageTickParams,
-	candidateBodies: readonly CandidateBody[],
-): string {
-	const systemPrompt = buildSystemPrompt({
-		userPromptSuffix: params.systemPrompt,
-		maxPerTick: params.maxPerTick,
-		candidates: params.candidates,
-	});
-	const { text: batchSummary } = buildTickUserMessage(
-		params.candidates,
-		params.repos,
-	);
-	const bodies = candidateBodies
-		.map(({ candidate, body }) => {
-			return `<candidate-body id="${escapeXml(candidate.id)}">
-${escapeXml(body)}
-</candidate-body>`;
-		})
-		.join("\n\n");
-
-	return `${systemPrompt}
-
-${batchSummary}
-
-<candidate-bodies>
-${bodies}
-</candidate-bodies>
-
-Return ONLY a JSON object with this shape:
-{
-  "proposals": [
-    {
-      "candidateId": "candidate id",
-      "taskAnchor": "message id or issue/pr id",
-      "repoId": "repo id",
-      "title": "short title",
-      "branchName": "lowercase-hyphen-branch",
-      "planMessage": "markdown plan"
-    }
-  ],
-  "skips": [
-    { "candidateId": "candidate id", "reason": "one sentence" }
-  ],
-  "summary": "short summary"
-}
-
-Use at most ${Math.max(1, params.maxPerTick)} proposals. Include a candidate in "skips" only when the whole candidate has no actionable task.`;
-}
-
-export function parseCodexDecision(text: string): TriageDecision {
-	const jsonText = extractJsonObject(text);
-	const parsed = JSON.parse(jsonText) as Record<string, unknown>;
-	const proposals = Array.isArray(parsed.proposals)
-		? parsed.proposals
-				.map(coerceProposal)
-				.filter((p): p is TriageProposal => p !== null)
-		: [];
-	const skips = Array.isArray(parsed.skips)
-		? parsed.skips
-				.map(coerceSkip)
-				.filter((s): s is { candidateId: string; reason: string } => s !== null)
-		: [];
-	const summary =
-		typeof parsed.summary === "string" && parsed.summary.trim().length > 0
-			? parsed.summary.trim()
-			: null;
-	return { proposals, skips, summary };
-}
-
-function extractJsonObject(text: string): string {
-	const fenced = text.match(/```(?:json)?\s*([\s\S]*?)```/i);
-	const source = fenced?.[1] ?? text;
-	const start = source.indexOf("{");
-	const end = source.lastIndexOf("}");
-	if (start === -1 || end === -1 || end <= start) {
-		throw new Error("Codex triage response did not contain JSON");
-	}
-	return source.slice(start, end + 1);
-}
-
-function coerceProposal(value: unknown): TriageProposal | null {
-	if (!value || typeof value !== "object") return null;
-	const obj = value as Record<string, unknown>;
-	const candidateId = stringField(obj, "candidateId");
-	const taskAnchor = stringField(obj, "taskAnchor");
-	const repoId = stringField(obj, "repoId");
-	const title = stringField(obj, "title");
-	const branchName = stringField(obj, "branchName");
-	const planMessage = stringField(obj, "planMessage");
-	if (
-		!candidateId ||
-		!taskAnchor ||
-		!repoId ||
-		!title ||
-		!branchName ||
-		!planMessage
-	) {
-		return null;
-	}
-	return { candidateId, taskAnchor, repoId, title, branchName, planMessage };
-}
-
-function coerceSkip(
-	value: unknown,
-): { candidateId: string; reason: string } | null {
-	if (!value || typeof value !== "object") return null;
-	const obj = value as Record<string, unknown>;
-	const candidateId = stringField(obj, "candidateId");
-	const reason = stringField(obj, "reason");
-	if (!candidateId || !reason) return null;
-	return { candidateId, reason };
-}
-
-function stringField(obj: Record<string, unknown>, key: string): string | null {
-	const value = obj[key];
-	if (typeof value !== "string") return null;
-	const trimmed = value.trim();
-	return trimmed.length > 0 ? trimmed : null;
-}
-
 async function persistSkips(
 	skips: readonly { candidateId: string; reason: string }[],
 ): Promise<void> {
@@ -431,22 +250,4 @@ function getNestedString(obj: unknown, ...keys: string[]): string | null {
 		current = (current as Record<string, unknown>)[key];
 	}
 	return typeof current === "string" ? current : null;
-}
-
-function escapeXml(value: string): string {
-	return value
-		.replace(/&/g, "&amp;")
-		.replace(/</g, "&lt;")
-		.replace(/>/g, "&gt;");
-}
-
-function safeFilePart(value: string): string {
-	return value.replace(/[^a-zA-Z0-9._-]+/g, "-").slice(0, 80) || "candidate";
-}
-
-function extensionForMime(mime: string): string {
-	if (mime === "image/jpeg") return "jpg";
-	if (mime === "image/webp") return "webp";
-	if (mime === "image/gif") return "gif";
-	return "png";
 }
